requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
structlog==25.4.0
faiss-cpu==1.11.0.post1
numpy==2.3.2
sentence-transformers==5.1.0
langdetect==1.0.9
stripe==11.6.0
# Whisper fallback transcription (local, no API key needed)
faster-whisper>=1.1.0
# Hybrid speaker diarization — requires PYANNOTE_AUTH_TOKEN env var
# and accepting the model terms at https://hf.co/pyannote/speaker-diarization-3.1
pyannote.audio>=3.3.2
